Added Make dropdown.

+++ VehicleManager.Business/IUserVehicleManager.cs
@@ -6,9 +6,9 @@ namespace VehicleManager.Business
     public interface IUserVehicleManager
     {
         IEnumerable<UserVehicle> GetVehicles(string userIdentifier);
-        void AddVehicle(string userIdentity, string make, uint mpg);
+        void AddVehicle(string userIdentity, string make, byte mpg);
         void DeleteVehicle(Guid id);
         UserVehicle FindVehicle(Guid id);
-        void UpdateVehicleMpg(Guid id, uint mpg);
+        void UpdateVehicleMpg(Guid id, byte mpg);
     }
 }
