+++ VehicleManager.Business/UserVehicle.cs
@@ -6,7 +6,7 @@ namespace VehicleManager.Business
     {
         public Guid Id { get; set; }
         public string UserId { get; set; }
-        public uint Mpg { get; set; }
+        public byte Mpg { get; set; }
         public VehicleMake Make { get; set; }
     }
 }
